refactor: adjust UI / view folder structure
