+++ lib/ui/views/survey/answer/answer_item_checkbox.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:prapare/models/fhir_questionnaire/survey/export.dart';
+import 'package:prapare/ui/views/survey/question/question_item_controller.dart';
 
 import 'answer_item_checkbox_controller.dart';
 import 'answer_title.dart';
-import 'question_item_controller.dart';
 
 class AnswerItemCheckbox extends StatelessWidget {
   const AnswerItemCheckbox(
